requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# LangChain & LangGraph
langgraph==0.2.50
langchain==0.3.9
langchain-openai==0.2.9
langchain-core==0.3.21

# OpenAI
openai==1.57.0

# Database & Analytics
duckdb==1.1.3
pyarrow==18.1.0
pandas==2.2.3

# Excel Processing
pyxlsb==1.0.10
openpyxl==3.1.5

# Firebase
firebase-admin==6.6.0

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Environment & Configuration
python-dotenv==1.0.1
pyyaml==6.0.2

# SQL Parsing & Validation
sqlparse==0.5.3

# Utilities
httpx==0.28.1
aiofiles==24.1.0

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
black==24.10.0
ruff==0.8.2
